Solving Ordinary Differential Equations to Understand Luminescence

Details

A collection of function to simulate luminescence signals in the mineral quartz based on published models.

Authors

Johannes Friedrich	University of Bayreuth, Germany

Sebastian Kreutzer	Geography & Earth Sciences, Aberystwyth University, United Kingdom

Christoph Schmidt	University of Bayreuth, Germany

Supervisor

Christoph Schmidt, University of Bayreuth, Germany
Sebastian Kreutzer,Geography & Earth Sciences, Aberystwyth University, United Kingdom #' Project source code repository

https://github.com/R-Lum/RLumModel

Related package projects

http://r-luminescence.de

https://CRAN.r-project.org/package=Luminescence

https://CRAN.r-project.org/package=RLumShiny

Package maintainer

Johannes Friedrich, University of Bayreuth, Germany
johannes.friedrich@posteo.de

Acknowledgement

The work of Johannes Friedrich is gratefully supported by the DFG in framework of the project 'Modelling quartz luminescence signal dynamics relevant for dating and dosimetry' (SCHM 305114-1).