Set parameters for Different Quartz Luminescence Models

This function provides all necessary model parameters to the calculation of the ODEs.

.set_pars(model)

Arguments

model

character (required): set model to be used. Available models are: "Bailey2001", "Bailey2002", "Bailey2004", "Pagonis2007", "Pagonis2008", "Friedrich2017", "Friedrich2018", "Peng2022". If model is indeed missing, the list of allowed keywords is returned.

Value

This function returns a list with all necessary parameters for the used model. Returns vector of allowed keywords if model is missing.

Details

The common model parameters are:

N: concentrations of electron/hole traps in cm^(-3) E: depth of the electron/hole trap in eV s: frequency factor in s^(-1) A: conduction band to electron/hole trap transition probability in s^(-1) B: valence band to hole trap transition probability in s^(-1) Th: photo-ionisation cross-section in s^(-1) E_th: thermal assistance energy in (eV) n: concentrations of electron/hole traps after sample history in cm^(-3)

Note

n are the saved concentrations of the last step of the sample history of the used model. They will be loaded, if simulate_sample_history = FALSE in model_LuminescenceSignals is chosen.

The order of the energy-band-levels is sometimes in an different order than in the original model. This was necessary, because in the simulations the luminescence centre always has to be the last entry in every parameter. Another reason was the clear division between electron traps and hole centres. When a user wants to create his/her own parameter sets he/she only has to take care that the luminescence centre is the last entry in every vector and that the first entries are the electron traps and the last entries the hole centres.

Function version

0.1.3

How to cite

Friedrich, J., 2022. .set_pars(): Set parameters for Different Quartz Luminescence Models. Function version 0.1.3. In: Friedrich, J., Kreutzer, S., Schmidt, C., 2022. RLumModel: Solving Ordinary Differential Equations to Understand Luminescence. R package version 0.2.10. https://CRAN.R-project.org/package=RLumModel

References

Bailey, R.M., 2001. Towards a general kinetic model for optically and thermally stimulated luminescence of quartz. Radiation Measurements 33, 17-45.

Bailey, R.M., 2002. Simulations of variability in the luminescence characteristics of natural quartz and its implications for estimates of absorbed dose. Radiation Protection Dosimetry 100, 33-38.

Bailey, R.M., 2004. Paper I-simulation of dose absorption in quartz over geological timescales and its implications for the precision and accuracy of optical dating. Radiation Measurements 38, 299-310.

Friedrich, J., Pagonis, V., Chen, R., Kreutzer, S., Schmidt, C., 2017: Quartz radiofluorescence: a modelling approach. Journal of Luminescence 186, 318-325.

Pagonis, V., Chen, R., Wintle, A.G., 2007. Modelling thermal transfer in optically stimulated luminescence of quartz. Journal of Physics D: Applied Physics 40, 998-1006.

Pagonis, V., Wintle, A.G., Chen, R., Wang, X.L., 2008. A theoretical model for a new dating protocol for quartz based on thermally transferred OSL (TT-OSL). Radiation Measurements 43, 704-708.

Peng, J., Wang, X., Adamiec, G., 2022. The build-up of the laboratory-generated dose-response curve and underestimation of equivalent dose for quartz OSL in the high dose region: A critical modelling study. Quaternary Geochronology 67, 101231.

Author

Johannes Friedrich, University of Bayreuth (Germany),

Examples

pars <- .set_pars("Bailey2001")