Parse a Risoe SEQ-file to a sequence neccessary for simulating quartz luminescence

A SEQ-file created by the Risoe Sequence Editor can be imported to simulate the sequence written in the sequence editor.

read_SEQ2R(file, lab.dose_rate = 1, txtProgressBar = TRUE)

Arguments

file
character (required): a *.seq file created by the Risoe Sequence Editor

lab.dose_rate
character (with default): set the dose rate of the radiation source in the laboratory Gy/s. Default: 1 Gy/s

txtProgressBar
logical (with default): enables or disables the txtProgressBar for a visuell control of the progress. Default: txtProgressBar = TRUE

Value

This function returns a list with the parsed *.seq file and the required steps for model_LuminescenceSignals.

Details

Supported versions: Supppored and tested: version 4.36.

Function version

0.1.0

How to cite

Friedrich, J., 2021. read_SEQ2R(): Parse a Risoe SEQ-file to a sequence neccessary for simulating quartz luminescence. Function version 0.1.0. In: Friedrich, J., Kreutzer, S., Schmidt, C., 2021. RLumModel: Solving Ordinary Differential Equations to Understand Luminescence. R package version 0.2.8. https://CRAN.R-project.org/package=RLumModel

References

Riso: Sequence Editor User Manual. Available at: http://www.nutech.dtu.dk/english/-/media/Andre_Universitetsenheder/Nutech/Produkter SequenceEditor.ashx?la=da

See also

Author

Johannes Friedrich, University of Bayreuth (Germany),

Examples

##search "example_SAR_cycle.SEQ" in "extdata" in package "RLumModel"
path <- system.file("extdata", "example_SAR_cycle.SEQ", package="RLumModel")

sequence <- read_SEQ2R(file = path, txtProgressBar = FALSE)